fix: product and purchase integration (#23)

+++ internal/model/purchase.go
@@ -7,7 +7,7 @@ type ProductCart struct {
 	Name             string    `json:"name"`
 	Category         string    `json:"category"`
 	Qty              int       `json:"qty"`   // Quantity before bought
-	Price            int64     `json:"price"` // Price per item
+	Price            float64   `json:"price"` // Price per item
 	SKU              string    `json:"sku"`
 	FileID           string    `json:"fileId"`
 	FileURI          string    `json:"fileUri"`          // Related file URI
@@ -17,17 +17,17 @@ type ProductCart struct {
 }
 
 type CartPaymentDetail struct {
-	BankAccountName   string `json:"bankAccountName"`
-	BankAccountHolder string `json:"bankAccountHolder"`
-	BankAccountNumber int64  `json:"bankAccountNumber"`
-	TotalPrice        int64  `json:"totalPrice"` // Total for this seller
+	BankAccountName   string  `json:"bankAccountName"`
+	BankAccountHolder string  `json:"bankAccountHolder"`
+	BankAccountNumber int64   `json:"bankAccountNumber"`
+	TotalPrice        float64 `json:"totalPrice"` // Total for this seller
 }
 
 type StoreCart struct {
-	TotalPrice          int64  `json:"totalPrice"`
-	SenderName          string `json:"senderName"`
-	SenderContactType   string `json:"senderContactType"`
-	SenderContactDetail string `json:"senderContactDetail"`
+	TotalPrice          float64 `json:"totalPrice"`
+	SenderName          string  `json:"senderName"`
+	SenderContactType   string  `json:"senderContactType"`
+	SenderContactDetail string  `json:"senderContactDetail"`
 }
 
 type StoreCartItems struct {
@@ -35,7 +35,7 @@ type StoreCartItems struct {
 	SellerID  int64
 	ProductID int64
 	Qty       int
-	Price     int64
+	Price     float64
 }
 
 // type PurchaseCartStore struct {
